project_01_enterprise_sales_data_warehouse/streaming_kafka/requirements.txt
kafka-python==2.0.2
python-dateutil==2.9.0.post0
